use better index types for tabular rows, columns and cells than int
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@9334 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/controllers/ControlTabularCreate.h
@@ -34,7 +34,7 @@ public:
 	virtual bool isBufferDependent() const { return true; }
 
 	///
-	typedef std::pair<unsigned int, unsigned int> rowsCols;
+	typedef std::pair<size_t, size_t> rowsCols;
 	///
 	rowsCols & params() { return params_; }
